Complete move to index folder
