+++ frontend/src/javascripts/components/index/guest-only.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import { Route, Switch, Redirect } from 'react-router-dom'
 import { connect } from 'react-redux'
-import TopPage from './top-page'
+import TopPage from '../guests/top-page'
 
 class GuestOnly extends Component {
   render() {
